pyproject.toml
[tool.poetry]
name = "polympics-vote"
version = "0.1.0"
description = "An interaction-based Discord bot for artwork voting in the Polympics."
authors = ["Artemis <artemis@artemisdev.xyz>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/polympics/vote-bot"

[tool.poetry.dependencies]
python = "^3.9"
polympics = "^0.6.2"
dslash = "^0.1.1"
"discord.py" = {git = "https://github.com/rapptz/discord.py.git"}
peewee = "^3.14.4"
psycopg2 = "^2.9.1"

[tool.poetry.dev-dependencies]
poethepoet = "^0.10.0"
black = "^21.7b0"
isort = "^5.9.3"
flake8 = "^3.9.2"
flake8-annotations = "^2.6.2"
flake8-bugbear = "^21.4.3"
flake8-comprehensions = "^3.5.0"
flake8-datetimez = "^20.10.0"
flake8-docstrings = "^1.6.0"
flake8-mutable = "^1.2.0"
flake8-raise = "^0.0.5"
flake8-string-format = "^0.3.0"
flake8-todo = "^0.7"

[tool.isort]
profile = "black"
src_paths = ["dslash"]

[tool.poe.tasks]
_format = "black . -q"
_format_imports = "isort . -q"
lint = "flake8 ."
format = ["_format_imports", "_format", "lint"]
bot = "python3 -m vote-bot"
